feat: support watcher for vite plugin adapter
