+++ packages/core/package.json
@@ -76,6 +76,7 @@
     "@swc/helpers": "^0.5.0",
     "boxen": "^7.0.1",
     "chalk": "^5.2.0",
+    "chokidar": "^3.5.3",
     "dotenv": "^16.3.1",
     "dotenv-expand": "^10.0.0",
     "execa": "^7.1.1",
