+++ packages/core/src/plugin/js/index.ts
@@ -1,82 +1,18 @@
 import merge from 'lodash.merge';
 import { Config } from '../../../binding/index.js';
 import {
-  isArray,
-  rustPluginResolver,
   type JsPlugin,
   type UserConfig,
-  convertPlugin,
   normalizeDevServerOptions
 } from '../../index.js';
-import { VITE_PLUGIN_DEFAULT_MODULE_TYPE, isObject } from './utils.js';
+import { VITE_PLUGIN_DEFAULT_MODULE_TYPE } from './utils.js';
 import { VitePluginAdapter } from './vite-plugin-adapter.js';
 import { existsSync, readFileSync } from 'node:fs';
 
 // export * from './jsPluginAdapter.js';
 export { VitePluginAdapter } from './vite-plugin-adapter.js';
 
-/**
- * resolvePlugins split / jsPlugins / rustPlugins
- * @param config
- */
-export async function resolveAllPlugins(
-  finalConfig: Config['config'],
-  userConfig: UserConfig
-) {
-  const plugins = userConfig.plugins ?? [];
-  const vitePlugins = userConfig.vitePlugins ?? [];
-
-  if (!plugins.length && !vitePlugins?.length) {
-    return {
-      rustPlugins: [],
-      jsPlugins: [],
-      finalConfig
-    };
-  }
-
-  const rustPlugins = [];
-
-  const jsPlugins: JsPlugin[] = handleVitePlugins(
-    vitePlugins,
-    userConfig,
-    finalConfig
-  );
-
-  for (const plugin of plugins) {
-    if (
-      typeof plugin === 'string' ||
-      (isArray(plugin) && typeof plugin[0] === 'string')
-    ) {
-      rustPlugins.push(
-        await rustPluginResolver(plugin as string, finalConfig.root)
-      );
-    } else if (isObject(plugin)) {
-      convertPlugin(plugin as unknown as JsPlugin);
-      jsPlugins.push(plugin as unknown as JsPlugin);
-    } else if (isArray(plugin)) {
-      for (const pluginNestItem of plugin as JsPlugin[]) {
-        convertPlugin(pluginNestItem as JsPlugin);
-        jsPlugins.push(pluginNestItem as JsPlugin);
-      }
-    } else {
-      throw new Error(
-        `plugin ${plugin} is not supported, Please pass the correct plugin type`
-      );
-    }
-  }
-  // call user config hooks
-  for (const jsPlugin of jsPlugins) {
-    finalConfig = (await jsPlugin.config?.(finalConfig)) ?? finalConfig;
-  }
-
-  return {
-    rustPlugins,
-    jsPlugins,
-    finalConfig
-  };
-}
-
-function handleVitePlugins(
+export function handleVitePlugins(
   vitePlugins: (object | (() => { vitePlugin: any; filters: string[] }))[],
   userConfig: UserConfig,
   finalConfig: Config['config']
@@ -138,4 +74,23 @@ function handleVitePlugins(
   }
 
   return jsPlugins;
+}
+
+export function convertPlugin(plugin: JsPlugin): void {
+  if (
+    plugin.transform &&
+    !plugin.transform.filters?.moduleTypes &&
+    !plugin.transform.filters?.resolvedPaths
+  ) {
+    throw new Error(
+      `transform hook of plugin ${plugin.name} must have at least one filter(like moduleTypes or resolvedPaths)`
+    );
+  }
+  if (plugin.transform) {
+    if (!plugin.transform.filters.moduleTypes) {
+      plugin.transform.filters.moduleTypes = [];
+    } else if (!plugin.transform.filters.resolvedPaths) {
+      plugin.transform.filters.resolvedPaths = [];
+    }
+  }
 }
